requirements.txt
Pillow>=10.0.0
cryptography>=41.0.0